Fixed Visual Studio warnings
(cherry picked from commit acdd059f9b6551f93a87f5b6ca4695446c9c3d84)

+++ src/codecs/music_gme.c
@@ -238,7 +238,7 @@ static void *GME_CreateFromRW(struct SDL_RWops *src, SDL_bool freesrc)
     SDL_RWseek(src, 0, SDL_RW_SEEK_SET);
     mem = SDL_LoadFile_RW(src, &size, SDL_FALSE);
     if (mem) {
-        err = gme.gme_open_data(mem, size, &music->game_emu, music_spec.freq);
+        err = gme.gme_open_data(mem, (long)size, &music->game_emu, music_spec.freq);
         SDL_free(mem);
         if (err != 0) {
             GME_Delete(music);
@@ -312,13 +312,13 @@ static int GME_GetSome(void *context, void *data, int bytes, SDL_bool *done)
         return 0;
     }
 
-    err = gme.gme_play(music->game_emu, (music->buffer_size / 2), (short*)music->buffer);
+    err = gme.gme_play(music->game_emu, (int)(music->buffer_size / 2), (short*)music->buffer);
     if (err != NULL) {
         Mix_SetError("GME: %s", err);
         return 0;
     }
 
-    if (SDL_PutAudioStreamData(music->stream, music->buffer, music->buffer_size) < 0) {
+    if (SDL_PutAudioStreamData(music->stream, music->buffer, (int)music->buffer_size) < 0) {
         return -1;
     }
     return 0;
